backend/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
pdfminer.six
python-docx
pandas
python-multipart
